:art: 로그인 관련 로직 :recycle: 이미지 로직 리펙토링
