+++ src/hooks/useHandleImage.ts
@@ -1,5 +1,5 @@
 import { RequestError } from '@/type'
-import { useCallback, useState } from 'react'
+import { ChangeEvent, DragEvent, useCallback, useState } from 'react'
 import { UseMutationOptions } from 'react-query'
 import usePostImage from './usePostImage'
 
@@ -12,6 +12,7 @@ type UseMutationOption = Omit<
 
 const useHandleImage = (options?: UseMutationOption) => {
   const [name, setName] = useState<string>()
+  const [isDragging, setIsDragging] = useState(false)
   const { mutate, reset, ...rest } = usePostImage({
     ...options,
     onSettled: () => {
@@ -19,11 +20,30 @@ const useHandleImage = (options?: UseMutationOption) => {
     },
   })
 
-  const handleChangeImage = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
+  const handleDragIn = useCallback((e: DragEvent<HTMLLabelElement>) => {
+    e.preventDefault()
+    e.stopPropagation()
+  }, [])
+
+  const handleDragOut = useCallback((e: DragEvent<HTMLLabelElement>) => {
+    e.preventDefault()
+    e.stopPropagation()
+    setIsDragging(false)
+  }, [])
+
+  const handleDragOver = useCallback((e: DragEvent<HTMLLabelElement>) => {
+    e.preventDefault()
+    e.stopPropagation()
+    if (e.dataTransfer?.files) {
+      setIsDragging(true)
+    }
+  }, [])
+
+  const handleChangeDropImage = useCallback((e: DragEvent<HTMLLabelElement>) => {
     e.preventDefault()
 
-    if (e.target.files) {
-      const file: File = e.target.files[0]
+    if (e.dataTransfer?.files) {
+      const file = e.dataTransfer?.files[0]
       const formData = new FormData()
       formData.append('file', file)
       setName(file.name)
@@ -31,7 +51,32 @@ const useHandleImage = (options?: UseMutationOption) => {
     }
   }, [])
 
+  const handleChangeImage = useCallback((e: ChangeEvent<HTMLInputElement>) => {
+    e.preventDefault()
+
+    if (e.target?.files) {
+      const file = e.target?.files[0]
+      const formData = new FormData()
+      formData.append('file', file)
+      setName(file.name)
+      mutate(formData)
+    }
+  }, [])
+
+  const handleDrop = useCallback((e: DragEvent<HTMLLabelElement>) => {
+    e.preventDefault()
+    e.stopPropagation()
+
+    setIsDragging(false)
+    handleChangeDropImage(e)
+  }, [])
+
   return {
+    isDragging,
+    handleDragIn,
+    handleDragOver,
+    handleDrop,
+    handleDragOut,
     handleChangeImage,
     name,
     ...rest,
